backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
strawberry-graphql[fastapi]
python-jose[cryptography]
pytest
pytest-asyncio